Cherry-picking the reversion of the change that disabled Scanning
The changes in this commit represent the REVERSION of the changes which initially disabled the scanning feature. I had some conflicts to resolve, and it was clear that the underlying codebase has changed a fair bit in the year+ since these changes were first applies. I tested in-game and there are many errors causing the TSM addon to not work. Further testing will be required